# apps/shop/api/openapi.yaml
openapi: 3.1.0
info:
  title: QA 숍 API
  version: 1.0.0
  description: |
    실습용 쇼핑몰 API. 이 문서가 API **계약의 원본**이다(spec-first).
    업무 규칙(금액 계산, 쿠폰, 배송, 상태 전이)은 SPEC.md 를 따른다.
    ⚠️ 의도적인 결함이 들어 있을 수 있는 실습용 서비스다. 공개 서버에 배포하지 말 것.
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:3000
...
paths:
...
  /api/members:
    post:
      tags: [members]
      summary: 회원 가입
      operationId: createMember
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewMember' }
      responses:
        '201':
          description: 가입됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/auth/login:
    post:
      tags: [members]
      summary: 로그인
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: 토큰 발급
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token: { type: string }
        '401': { $ref: '#/components/responses/Error' }
  /api/members/me:
    get:
      tags: [members]
      summary: 내 정보
      operationId: getMe
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '401': { $ref: '#/components/responses/Error' }
  /api/members/me/coupons:
    get:
      tags: [members]
      summary: 내 쿠폰 목록
      operationId: listMyCoupons
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OwnedCoupon' }
        '401': { $ref: '#/components/responses/Error' }
  /api/grades/evaluate:
    get:
      tags: [calculators]
      summary: 누적 구매액으로 등급 판정
      operationId: evaluateGrade
      parameters:
        - name: totalSpent
          in: query
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [totalSpent, grade, discountRate]
                properties:
                  totalSpent: { type: integer }
                  grade: { $ref: '#/components/schemas/Grade' }
                  discountRate: { type: integer, description: 상품 할인율(%) }
        '400': { $ref: '#/components/responses/Error' }
...
  /api/products/{id}:
    get:
      tags: [catalog]
      summary: 상품 상세
      operationId: getProduct
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message, details]
      properties:
        code: { type: string, examples: [VALIDATION_ERROR] }
        message: { type: string }
        details: { type: object }
    Zipcode:
      type: string
      pattern: '^[0-9]{5}$'
    Grade:
      type: string
      enum: [NORMAL, SILVER, GOLD, VIP]
    NewMember:
      type: object
      required: [name, email, password, zipcode]
      properties:
        name: { type: string, description: 공백 제거 후 2~20자 (한글 1글자 = 1자) }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 64 }
        zipcode: { $ref: '#/components/schemas/Zipcode' }
        address: { type: string }
    Member:
      type: object
      required: [id, email, name, zipcode, address, grade, totalSpent]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        zipcode: { $ref: '#/components/schemas/Zipcode' }
        address: { type: string }
        grade: { $ref: '#/components/schemas/Grade' }
        totalSpent: { type: integer, minimum: 0 }
    OwnedCoupon:
      type: object
      required: [code, type, amount, rate, maxDiscount, minOrderAmount, validFrom, validUntil, usedAt]
      properties:
        code: { type: string }
        type: { type: string, enum: [FIXED, PERCENT] }
        amount: { type: [integer, 'null'] }
        rate: { type: [integer, 'null'] }
        maxDiscount: { type: [integer, 'null'] }
        minOrderAmount: { type: integer }
        validFrom: { type: string, format: date }
        validUntil: { type: string, format: date }
        usedAt: { type: [string, 'null'], format: date-time }
    Product:
      type: object
      required: [id, name, price, stock]
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: integer, minimum: 1 }
        stock: { type: integer, minimum: 0 }
...